+++ modules/apps/rate-limiting/keeper/keeper_test.go
@@ -10,6 +10,7 @@ import (
 	"github.com/cosmos/ibc-go/v10/modules/apps/rate-limiting/keeper"
 	ratelimittypes "github.com/cosmos/ibc-go/v10/modules/apps/rate-limiting/types"
 	ibctesting "github.com/cosmos/ibc-go/v10/testing"
+	ibcmock "github.com/cosmos/ibc-go/v10/testing/mock"
 )
 
 type KeeperTestSuite struct {
@@ -44,6 +45,22 @@ func (s *KeeperTestSuite) TestNewKeeper() {
 			instantiateFn: func() {
 				keeper.NewKeeper(
 					s.chainA.GetSimApp().AppCodec(),
+					s.chainA.GetSimApp().AccountKeeper.AddressCodec(),
+					runtime.NewKVStoreService(s.chainA.GetSimApp().GetKey(ratelimittypes.StoreKey)),
+					s.chainA.GetSimApp().IBCKeeper.ChannelKeeper,
+					s.chainA.GetSimApp().IBCKeeper.ClientKeeper, // Add clientKeeper
+					s.chainA.GetSimApp().BankKeeper,
+					s.chainA.GetSimApp().ICAHostKeeper.GetAuthority(),
+				)
+			},
+			panicMsg: "",
+		},
+		{
+			name: "success: custom address codec",
+			instantiateFn: func() {
+				keeper.NewKeeper(
+					s.chainA.GetSimApp().AppCodec(),
+					ibcmock.TestAddressCodec{},
 					runtime.NewKVStoreService(s.chainA.GetSimApp().GetKey(ratelimittypes.StoreKey)),
 					s.chainA.GetSimApp().IBCKeeper.ChannelKeeper,
 					s.chainA.GetSimApp().IBCKeeper.ClientKeeper, // Add clientKeeper
@@ -58,6 +75,7 @@ func (s *KeeperTestSuite) TestNewKeeper() {
 			instantiateFn: func() {
 				keeper.NewKeeper(
 					s.chainA.GetSimApp().AppCodec(),
+					s.chainA.GetSimApp().AccountKeeper.AddressCodec(),
 					runtime.NewKVStoreService(s.chainA.GetSimApp().GetKey(ratelimittypes.StoreKey)),
 					s.chainA.GetSimApp().IBCKeeper.ChannelKeeper,
 					s.chainA.GetSimApp().IBCKeeper.ClientKeeper, // clientKeeper