+++ modules/apps/transfer/keeper/keeper.go
@@ -5,6 +5,7 @@ import (
 	"fmt"
 	"strings"
 
+	"cosmossdk.io/core/address"
 	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/log"
 	sdkmath "cosmossdk.io/math"
@@ -27,6 +28,7 @@ import (
 type Keeper struct {
 	storeService corestore.KVStoreService
 	cdc          codec.BinaryCodec
+	addressCodec address.Codec
 
 	ics4Wrapper   porttypes.ICS4Wrapper
 	channelKeeper types.ChannelKeeper
@@ -40,15 +42,7 @@ type Keeper struct {
 }
 
 // NewKeeper creates a new IBC transfer Keeper instance
-func NewKeeper(
-	cdc codec.BinaryCodec,
-	storeService corestore.KVStoreService,
-	channelKeeper types.ChannelKeeper,
-	msgRouter types.MessageRouter,
-	authKeeper types.AccountKeeper,
-	bankKeeper types.BankKeeper,
-	authority string,
-) *Keeper {
+func NewKeeper(cdc codec.BinaryCodec, addressCodec address.Codec, storeService corestore.KVStoreService, channelKeeper types.ChannelKeeper, msgRouter types.MessageRouter, authKeeper types.AccountKeeper, bankKeeper types.BankKeeper, authority string) *Keeper {
 	// ensure ibc transfer module account is set
 	if addr := authKeeper.GetModuleAddress(types.ModuleName); addr == nil {
 		panic(errors.New("the IBC transfer module account has not been set"))
@@ -60,6 +54,7 @@ func NewKeeper(
 
 	return &Keeper{
 		cdc:           cdc,
+		addressCodec:  addressCodec,
 		storeService:  storeService,
 		ics4Wrapper:   channelKeeper, // default ICS4Wrapper is the channel keeper
 		channelKeeper: channelKeeper,
@@ -85,6 +80,11 @@ func (k *Keeper) GetICS4Wrapper() porttypes.ICS4Wrapper {
 // GetAuthority returns the transfer module's authority.
 func (k *Keeper) GetAuthority() string {
 	return k.authority
+}
+
+// GetAddressCodec returns the address codec used by the keeper.
+func (k *Keeper) GetAddressCodec() address.Codec {
+	return k.addressCodec
 }
 
 // Logger returns a module-specific logger.