+++ modules/apps/transfer/keeper/relay_test.go
@@ -1,6 +1,7 @@
 package keeper_test
 
 import (
+	"encoding/hex"
 	"errors"
 	"fmt"
 	"strings"
@@ -47,19 +48,19 @@ func (s *KeeperTestSuite) TestSendTransfer() {
 		expError error
 	}{
 		{
-			"successful transfer of native token",
+			"success: transfer of native token",
 			func() {},
 			nil,
 		},
 		{
-			"successful transfer of native token with memo",
+			"success: transfer of native token with memo",
 			func() {
 				memo = "memo" //nolint:goconst
 			},
 			nil,
 		},
 		{
-			"successful transfer of IBC token",
+			"success: transfer of IBC token",
 			func() {
 				// send IBC token back to chainB
 				denom := types.NewDenom(ibctesting.TestCoin.Denom, types.NewHop(path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID))
@@ -70,7 +71,7 @@ func (s *KeeperTestSuite) TestSendTransfer() {
 			nil,
 		},
 		{
-			"successful transfer of IBC token with memo",
+			"success: transfer of IBC token with memo",
 			func() {
 				// send IBC token back to chainB
 				denom := types.NewDenom(ibctesting.TestCoin.Denom, types.NewHop(path.EndpointA.ChannelConfig.PortID, path.EndpointA.ChannelID))
@@ -82,7 +83,7 @@ func (s *KeeperTestSuite) TestSendTransfer() {
 			nil,
 		},
 		{
-			"successful transfer of entire balance",
+			"success: transfer of entire balance",
 			func() {
 				coin = sdk.NewCoin(coin.Denom, types.UnboundedSpendLimit())
 				var ok bool
@@ -92,7 +93,7 @@ func (s *KeeperTestSuite) TestSendTransfer() {
 			nil,
 		},
 		{
-			"successful transfer of entire spendable balance with vesting account",
+			"success: transfer of entire spendable balance with vesting account",
 			func() {
 				// create vesting account
 				vestingAccPrivKey := secp256k1.GenPrivKey()
@@ -325,14 +326,24 @@ func (s *KeeperTestSuite) TestOnRecvPacket_ReceiverIsNotSource() {
 		expError error
 	}{
 		{
-			"successful receive",
+			"success: receive",
 			func() {},
 			nil,
 		},
 		{
-			"successful receive with memo",
+			"success: receive with memo",
 			func() {
 				packetData.Memo = "memo"
+			},
+			nil,
+		},
+		{
+			"success: receive with hex receiver address",
+			func() {
+				s.chainB.GetSimApp().TransferKeeper.SetAddressCodec(ibcmock.TestAddressCodec{})
+
+				receiver := sdk.MustAccAddressFromBech32(packetData.Receiver)
+				packetData.Receiver = hex.EncodeToString(receiver.Bytes())
 			},
 			nil,
 		},